Bug 23344 - Branch Transfer Limits - Warn if sysprefs are not set
The admin-menu branch_transfer_limits.pl is available but doesn't say anything about sysprefs that need setting.
Testing and finding out the limits don't work is confusing.

It would be more intuitive to just use the feature in the background without having to mess with any extra sysprefs.

Or set the aforementioned sysprefs in the branch_transfer_limits.pl -page.

This patch adds a simple warning if syspref 'UseBranchTransferLimits' is not activated.

Sponsored-by: The National Library of Finland