Bug 37744 - MARC21slimUtils.xslt show-series has a wrong search index 'au' for 810$a
Using Elasticsearch

810$a is not indexed in Elastic search mappings as an 'author'-index -field.
810 is mapped to the 'author-name-corporate'-index.
'author-name-corporate'-index is the correct index for 810-field.

TEST PLAN:
- Create a record with Fields 810$a and 810$t
- A Series-link is generated in the biblio detail.pl and search results XSLT-views.
- The link has a bad search query: (title-series:"Value from 810$t") (author:"Value from 810$a")
-- 'author'-search doesnt find anything as the 810$a is not indexed to 'author'-index.
- This patch changes the 'author'-search to 'author-name-corporate'-search, and the search returns results.

Alternatively:
It might be desirable to add 810$a to the 'author'-index.
There is a specific search for 'author-name-corporate' using the advanced search.

Related to:
Bug XXXX - MARC21 810$t is used in XSLT templates but is not editable by default.